pyproject.toml
[tool.poetry]
name = "med_list2"
version = "0.1.0"
description = "Telegram bot with unworking drugs"
authors = ["Voronwe <Voronwe.S@protonmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
psycopg2-binary = "^2.9.10"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
lxml = "^5.3.0"
python-telegram-bot = "^21.6"
asgiref = "^3.8.1"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
